backend/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
supabase==2.5.1
PyJWT==2.8.0
werkzeug==3.0.3
